PROTON-1410 Moving ThreadLocal logic for Decoders on Message to its own class
I need to reuse the same Decoder used by the MessageImpl when creating my own special Message Type.

https://issues.apache.org/jira/browse/PROTON-1410
